sdk/tests/conformance2/extensions/required-extensions.html

<!--
Copyright (c) 2021 The Khronos Group Inc.
Use of this source code is governed by an MIT-style license that can be
found in the LICENSE.txt file.
-->

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<link rel="stylesheet" href="../../resources/js-test-style.css"/>
<script src="../../js/js-test-pre.js"></script>
<script src="../../js/webgl-test-utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>

<script>
"use strict";
const wtu = WebGLTestUtils;
let gl;

description("Ensure that required extensions are supported");

shouldBeNonNull("gl = wtu.create3DContext(undefined, undefined, 2)");

const supportedExts = gl.getSupportedExtensions();

function hasExt(name) {
    return supportedExts.indexOf(name) >= 0;
}

shouldBeTrue("hasExt('WEBGL_compressed_texture_s3tc') || hasExt('WEBGL_compressed_texture_etc')");

debug("")
var successfullyParsed = true;
</script>

<script src="../../js/js-test-post.js"></script>
</body>
</html>
